Restructured extract-imga.py
